Added features to select priv level (U or M only)
Many embedded RISC-V cores implement M and U modes. With the way the
library was previously written, it was impossible to use this library
as-is in any privelege level besides M mode because of interrupt
suspension. This commit adds features for selecting a mode, which will
cause a version of the functions in export.rs to be compiled with the
correct interrupt suspension, if at all. One of the features must be
selected, otherwise a compiler error will be thrown.

Currently there is no support for Supervisor mode as there is no support
for it in `riscv::interrupt`.